+++ 2023/day_01/second_part.cpp
@@ -11,7 +11,7 @@ std::string digits[9] = {
 
 int longestPrefix(const std::string& line, const std::string& substring, int start) {
     int j = 0;
-    while (j < substring.size() && start + j < line.size() && substring[j] == line[start + j]) {
+    while (j < (int)substring.size() && start + j < (int)line.size() && substring[j] == line[start + j]) {
         j++;
     }
     return j;
@@ -21,7 +21,7 @@ int digitChecker(const std::string& line, int i) {
     if (isdigit(line[i])) return line[i] - '0';
     for (int d = 1; d <= 9; d++) {
         int j = longestPrefix(line, digits[d - 1], i);
-        if (j == digits[d - 1].size()) return d;
+        if (j == (int)digits[d - 1].size()) return d;
     }
     return -1;
 }
