+++ 2023/day_06/second_part.cpp
@@ -9,9 +9,9 @@ int solver() {
 
     ll time = 0;
     int i = 0;
-    while (i < lines[0].size()) {
+    while (i < (int)lines[0].size()) {
         utils::skipUntil(lines[0], i, [](char c){return (bool)isdigit(c);});
-        while (i < lines[0].size() && isdigit(lines[0][i])) {
+        while (i < (int)lines[0].size() && isdigit(lines[0][i])) {
             time = time * 10 + (lines[0][i] - '0');
             i++;
         }
@@ -19,9 +19,9 @@ int solver() {
 
     ll distance = 0;
     int j = 0;
-    while (j < lines[1].size()) {
+    while (j < (int)lines[1].size()) {
         utils::skipUntil(lines[1], j, [](char c){return (bool)isdigit(c);});
-        while (j < lines[1].size() && isdigit(lines[1][j])) {
+        while (j < (int)lines[1].size() && isdigit(lines[1][j])) {
             distance = distance * 10 + (lines[1][j] - '0');
             j++;
         }
